+++ src/pages/event/WriteEventFormPage/WriteEventFormPage.tsx
@@ -1,6 +1,6 @@
 import FormCheckOptions from '@/components/FormCheckOptions/FormCheckOptions';
 import FormInformationEditor from '@/components/FormInformationEditor/FormInformationEditor';
-import FormOptionButton from '@/components/FormOptionButton/FormOptionButton';
+import FormOptionButton from '@/components/FormOptionButtons/FormOptionButtons';
 import FormOptionDropdown from '@/components/FormOptionDropdown/FormOptionDropdown';
 import FormOptions from '@/components/FormOptions/FormOptions';
 import { EventType } from '@/types/event';
